Added loading of Sintons new Labview file output.
Current issue with its implimentation is that the labview output does
not provide a doping. This causes issues as a doping is required for
calculation of the excess carrier density. This value can be  calculated
from provided resistivity but that is not being done yet. It can also be
provided manually.